sanic-backend/requirements.txt
sanic>=22.12.0
sanic-ext>=22.12.0
lucidtech-synthetic